# Conflicts:
#	happybay/src/main/java/com/propra/happybay/Controller/UserController.java
#	happybay/src/main/resources/application-dev.properties
#	happybay/src/main/resources/application.properties

+++ happybay/src/main/java/com/propra/happybay/Model/GeraetMitReservationID.java
@@ -9,7 +9,7 @@ import javax.persistence.Id;
 @Entity
 public class GeraetMitReservationID {
     @Id
-    private int reservationID;
+    private Long reservationID;
     private Long geraetID;
 
 }
